added relation between user and files

+++ api.py
@@ -2,10 +2,12 @@ from hashlib import sha256
 from pathlib import Path
 
 from flask import request, send_from_directory
+from sqlalchemy import select
 from werkzeug.security import safe_join
 
-from app import app
+from app import app, db
 from auth import auth
+from models import File
 
 
 @app.route('/<filename>', methods = ['GET'])
@@ -28,8 +30,20 @@ def upload_file():
     # we use hash instead of origin filename so 
     # we can safely join the full path + no need to validate the extension
     filename = app.config['UPLOAD_FOLDER'] / filename_hash[:2] / filename_hash
+
+    if filename.exists():
+        return 'File with such name already exists', 400
+
     filename.parent.mkdir(parents=True, exist_ok=True)
     file.save(filename)
+    
+    db_file = File(
+        origin_name=file.filename,
+        hash=filename_hash,
+        user_id=auth.current_user().id
+    )
+    db.session.add(db_file)
+    db.session.commit()
 
     return filename_hash, 201
 
@@ -46,5 +60,12 @@ def delete_file(filename):
     if not filepath.exists() or not filepath.is_file():
         return 'File not found', 404
 
+    db_file = db.session.execute(select(File).where(File.hash == filepath.name)).first()[0]
+    if db_file.user_id != auth.current_user().id:
+        return "Forbidden", 403
+
     filepath.unlink()
+    db.session.delete(db_file)
+    db.session.commit()
+
     return 'Deleted', 200
